requirements.txt
flask
fugashi
jaconv
opencv-python
Pillow
pyclipper
scipy
shapely
torch
torchvision
transformers
unidic_lite
